refactor(cmd): Migrate all `runu` commands into `internal/cli/runu`
